Make Rack::Console object-oriented
It's a bit nicer to initialize a Rack::Console with the options hash and
then call `.start` on the instance. Additionally, move the option
parsing back into the `rack-console` binary. This allows people writing
their own CLIs around Rack::Console to mix and match the options they
actually want to support.

Signed-off-by: David Celis <me@davidcel.is>